Refactor cipher_suites off openssl_util.rs
